Drop support for PHP 8.1, Add support for PHP 8.5

+++ src/ItemList.php
@@ -8,6 +8,7 @@ use ArrayAccess;
 use Countable;
 use Laminas\Tag\Exception\InvalidArgumentException;
 use Laminas\Tag\Exception\OutOfBoundsException;
+use Override;
 use ReturnTypeWillChange;
 use SeekableIterator;
 
@@ -41,6 +42,7 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
      *
      * @return int
      */
+    #[Override]
     #[ReturnTypeWillChange]
     public function count()
     {
@@ -84,17 +86,17 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
             }
 
             // Calculate the thresholds
-            $steps      = count($values);
-            $delta      = ($maxWeight - $minWeight) / ($steps - 1);
+            $steps      = (float) count($values);
+            $delta      = ($maxWeight - $minWeight) / ($steps - 1.0);
             $thresholds = [];
 
             for ($i = 0; $i < $steps; $i++) {
-                $thresholds[$i] = floor(100 * log(($minWeight + $i * $delta) + 2));
+                $thresholds[$i] = floor(100.0 * log(($minWeight + (float) $i * $delta) + 2.0));
             }
 
             // Then assign the weight values
             foreach ($this->items as $item) {
-                $threshold = floor(100 * log($item->getWeight() + 2));
+                $threshold = floor(100.0 * log($item->getWeight() + 2.0));
 
                 for ($i = 0; $i < $steps; $i++) {
                     if ($threshold <= $thresholds[$i]) {
@@ -113,6 +115,7 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
      * @throws OutOfBoundsException When the seek position is invalid.
      * @return void
      */
+    #[Override]
     #[ReturnTypeWillChange]
     public function seek($index)
     {
@@ -134,6 +137,7 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
      *
      * @return TaggableInterface|false
      */
+    #[Override]
     #[ReturnTypeWillChange]
     public function current()
     {
@@ -145,6 +149,7 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
      *
      * @return TaggableInterface|false
      */
+    #[Override]
     #[ReturnTypeWillChange]
     public function next()
     {
@@ -156,10 +161,11 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
      *
      * @return int
      */
+    #[Override]
     #[ReturnTypeWillChange]
     public function key()
     {
-        return key($this->items);
+        return (int) key($this->items);
     }
 
     /**
@@ -167,6 +173,7 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
      *
      * @return bool
      */
+    #[Override]
     #[ReturnTypeWillChange]
     public function valid()
     {
@@ -178,6 +185,7 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
      *
      * @return void
      */
+    #[Override]
     #[ReturnTypeWillChange]
     public function rewind()
     {
@@ -190,6 +198,7 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
      * @param int $offset
      * @return bool
      */
+    #[Override]
     #[ReturnTypeWillChange]
     public function offsetExists($offset)
     {
@@ -202,6 +211,7 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
      * @param int $offset
      * @return TaggableInterface
      */
+    #[Override]
     #[ReturnTypeWillChange]
     public function offsetGet($offset)
     {
@@ -216,6 +226,7 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
      * @return void
      * @throws OutOfBoundsException When item does not implement Laminas\Tag\TaggableInterface.
      */
+    #[Override]
     #[ReturnTypeWillChange]
     public function offsetSet($offset, $value)
     {
@@ -238,6 +249,7 @@ class ItemList implements Countable, SeekableIterator, ArrayAccess
      * @param int $offset
      * @return void
      */
+    #[Override]
     #[ReturnTypeWillChange]
     public function offsetUnset($offset)
     {
